// src/@Types/IAgente.d.ts
export interface IAgente{
	uuid: string,
	displayName: string,
	description: string,
	developerName: string,
	characterTags: string,
	displayIcon: string,
	displayIconSmall: string,
	bustPortrait: string,
	fullPortrait: string,
	fullPortraitV2: string,
	killfeedPortrait: string,
	background: string,
	backgroundGradientColors: Array<string>,
	assetPath: string,
	isFullPortraitRightFacing: boolean,
	isPlayableCharacter: boolean,
	isAvailableForTest: boolean,
	isBaseContent: boolean,
	role: {
		uuid: string,
		displayName: string,
		description: string,
		displayIcon: string,
		assetPath: string,
	},
	abilities: [{
		  slot: string,
		  displayName: string,
		  description: string,
		  displayIcon: string,
	}],
	voiceline: string[],
  }